backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dateutil
